% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hourly_api.R
\docType{class}
\name{HourlyApi}
\alias{HourlyApi}
\title{Hourly operations}
\format{
An \code{R6Class} generator object
}
\description{
HourlyApi
}
\details{
AZMet API

REST API for the Arizona Meteorological Network (AZMet)

The version of the OpenAPI document: 1.0
Generated by: https://openapi-generator.tech
}
\section{Methods}{

\describe{
\strong{ v1_observations_hourly_get } \emph{  }
Pull hourly direct and derived observations for all stations over the last 24 hours.

\itemize{
\item \emph{ @returnType } \link{InlineResponse2002} \cr

\item On encountering errors, an error of subclass ApiException will be thrown.

\item status code : 200 | Pull hourly direct and derived observations for all stations over the last 24 hours.

\item return type : InlineResponse2002
\item response headers :

\tabular{ll}{
}
}

\strong{ v1_observations_hourly_station_id_date_time_start_get } \emph{ Return hourly direct and derived observations for a specific station and date and time. }
Pull hourly direct and derived observations for a specific station over the last 24 hours starting at a particular date and time.

\itemize{
\item \emph{ @param } station_id character
\item \emph{ @param } date_time_start character
\item \emph{ @returnType } \link{InlineResponse2003} \cr

\item On encountering errors, an error of subclass ApiException will be thrown.

\item status code : 200 | Pull hourly direct and derived observations for a specific station over the last 24 hours starting at a particular date and time.

\item return type : InlineResponse2003
\item response headers :

\tabular{ll}{
}
}

\strong{ v1_observations_hourly_station_id_date_time_start_time_interval_get } \emph{ Return hourly direct and derived observations for a specific station over a time interval from a date and time. }
Return hourly direct and derived observations for a specific station over a particular time interval starting at a particular date and time.

\itemize{
\item \emph{ @param } station_id character
\item \emph{ @param } date_time_start character
\item \emph{ @param } time_interval character
\item \emph{ @returnType } \link{InlineResponse2004} \cr

\item On encountering errors, an error of subclass ApiException will be thrown.

\item status code : 200 | The hourly direct and derived observations for a specific station over a time interval startig from a given date and time.

\item return type : InlineResponse2004
\item response headers :

\tabular{ll}{
}
}

\strong{ v1_observations_hourly_station_id_get } \emph{ Return hourly direct and derived observations for a specific station. }
Pull hourly direct and derived observations for a specific station over the last 24 hours.

\itemize{
\item \emph{ @param } station_id character
\item \emph{ @returnType } \link{InlineResponse2003} \cr

\item On encountering errors, an error of subclass ApiException will be thrown.

\item status code : 200 | The hourly direct and derived observations for a specific station over the last 24 hours.

\item return type : InlineResponse2003
\item response headers :

\tabular{ll}{
}
}

}
}

\examples{
\dontrun{
####################  v1_observations_hourly_get  ####################

library(azmet)

api_instance <- HourlyApi$new()

result <- tryCatch(
             
             # to save the result into a file, simply add the optional `data_file` parameter, e.g.
             # api_instance$v1_observations_hourly_get(data_file = "result.txt"),
             
             
             api_instance$v1_observations_hourly_get(),
             ApiException = function(ex) ex
          )
# In case of error, print the error object
if (!is.null(result$ApiException)) {
  print("Exception occurs when calling `v1_observations_hourly_get`:")
  dput(result$ApiException$toString())
  
  # error object
  dput(result$ApiException$error_object$toJSONString())
  
} else {
  # deserialized response object
  print("The response is ...")
  dput(result$toString())
}



####################  v1_observations_hourly_station_id_date_time_start_get  ####################

library(azmet)
var_station_id <- "station_id_example" # character | ID of station
var_date_time_start <- "date_time_start_example" # character | Date and time of data request start

#Return hourly direct and derived observations for a specific station and date and time.
api_instance <- HourlyApi$new()

result <- tryCatch(
             
             # to save the result into a file, simply add the optional `data_file` parameter, e.g.
             # api_instance$v1_observations_hourly_station_id_date_time_start_get(var_station_id, var_date_time_start, data_file = "result.txt"),
             
             
             api_instance$v1_observations_hourly_station_id_date_time_start_get(var_station_id, var_date_time_start),
             ApiException = function(ex) ex
          )
# In case of error, print the error object
if (!is.null(result$ApiException)) {
  print("Exception occurs when calling `v1_observations_hourly_station_id_date_time_start_get`:")
  dput(result$ApiException$toString())
  
  # error object
  dput(result$ApiException$error_object$toJSONString())
  
} else {
  # deserialized response object
  print("The response is ...")
  dput(result$toString())
}



####################  v1_observations_hourly_station_id_date_time_start_time_interval_get  ####################

library(azmet)
var_station_id <- "station_id_example" # character | ID of station
var_date_time_start <- "date_time_start_example" # character | Date and time of data request start
var_time_interval <- "time_interval_example" # character | ID of station

#Return hourly direct and derived observations for a specific station over a time interval from a date and time.
api_instance <- HourlyApi$new()

result <- tryCatch(
             
             # to save the result into a file, simply add the optional `data_file` parameter, e.g.
             # api_instance$v1_observations_hourly_station_id_date_time_start_time_interval_get(var_station_id, var_date_time_start, var_time_interval, data_file = "result.txt"),
             
             
             api_instance$v1_observations_hourly_station_id_date_time_start_time_interval_get(var_station_id, var_date_time_start, var_time_interval),
             ApiException = function(ex) ex
          )
# In case of error, print the error object
if (!is.null(result$ApiException)) {
  print("Exception occurs when calling `v1_observations_hourly_station_id_date_time_start_time_interval_get`:")
  dput(result$ApiException$toString())
  
  # error object
  dput(result$ApiException$error_object$toJSONString())
  
} else {
  # deserialized response object
  print("The response is ...")
  dput(result$toString())
}



####################  v1_observations_hourly_station_id_get  ####################

library(azmet)
var_station_id <- "station_id_example" # character | ID of station

#Return hourly direct and derived observations for a specific station.
api_instance <- HourlyApi$new()

result <- tryCatch(
             
             # to save the result into a file, simply add the optional `data_file` parameter, e.g.
             # api_instance$v1_observations_hourly_station_id_get(var_station_id, data_file = "result.txt"),
             
             
             api_instance$v1_observations_hourly_station_id_get(var_station_id),
             ApiException = function(ex) ex
          )
# In case of error, print the error object
if (!is.null(result$ApiException)) {
  print("Exception occurs when calling `v1_observations_hourly_station_id_get`:")
  dput(result$ApiException$toString())
  
  # error object
  dput(result$ApiException$error_object$toJSONString())
  
} else {
  # deserialized response object
  print("The response is ...")
  dput(result$toString())
}



}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{api_client}}{Handles the client-server communication.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-HourlyApi-new}{\code{HourlyApi$new()}}
\item \href{#method-HourlyApi-v1_observations_hourly_get}{\code{HourlyApi$v1_observations_hourly_get()}}
\item \href{#method-HourlyApi-v1_observations_hourly_get_with_http_info}{\code{HourlyApi$v1_observations_hourly_get_with_http_info()}}
\item \href{#method-HourlyApi-v1_observations_hourly_station_id_date_time_start_get}{\code{HourlyApi$v1_observations_hourly_station_id_date_time_start_get()}}
\item \href{#method-HourlyApi-v1_observations_hourly_station_id_date_time_start_get_with_http_info}{\code{HourlyApi$v1_observations_hourly_station_id_date_time_start_get_with_http_info()}}
\item \href{#method-HourlyApi-v1_observations_hourly_station_id_date_time_start_time_interval_get}{\code{HourlyApi$v1_observations_hourly_station_id_date_time_start_time_interval_get()}}
\item \href{#method-HourlyApi-v1_observations_hourly_station_id_date_time_start_time_interval_get_with_http_info}{\code{HourlyApi$v1_observations_hourly_station_id_date_time_start_time_interval_get_with_http_info()}}
\item \href{#method-HourlyApi-v1_observations_hourly_station_id_get}{\code{HourlyApi$v1_observations_hourly_station_id_get()}}
\item \href{#method-HourlyApi-v1_observations_hourly_station_id_get_with_http_info}{\code{HourlyApi$v1_observations_hourly_station_id_get_with_http_info()}}
\item \href{#method-HourlyApi-clone}{\code{HourlyApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HourlyApi-new"></a>}}
\if{latex}{\out{\hypertarget{method-HourlyApi-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new HourlyApi.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HourlyApi$new(api_client)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{api_client}}{An instance of API client.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HourlyApi-v1_observations_hourly_get"></a>}}
\if{latex}{\out{\hypertarget{method-HourlyApi-v1_observations_hourly_get}{}}}
\subsection{Method \code{v1_observations_hourly_get()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HourlyApi$v1_observations_hourly_get(data_file = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
InlineResponse2002
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HourlyApi-v1_observations_hourly_get_with_http_info"></a>}}
\if{latex}{\out{\hypertarget{method-HourlyApi-v1_observations_hourly_get_with_http_info}{}}}
\subsection{Method \code{v1_observations_hourly_get_with_http_info()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HourlyApi$v1_observations_hourly_get_with_http_info(data_file = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (InlineResponse2002) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HourlyApi-v1_observations_hourly_station_id_date_time_start_get"></a>}}
\if{latex}{\out{\hypertarget{method-HourlyApi-v1_observations_hourly_station_id_date_time_start_get}{}}}
\subsection{Method \code{v1_observations_hourly_station_id_date_time_start_get()}}{
Return hourly direct and derived observations for a specific station and date and time.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HourlyApi$v1_observations_hourly_station_id_date_time_start_get(
  station_id,
  date_time_start,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{station_id}}{ID of station}

\item{\code{date_time_start}}{Date and time of data request start}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
InlineResponse2003
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HourlyApi-v1_observations_hourly_station_id_date_time_start_get_with_http_info"></a>}}
\if{latex}{\out{\hypertarget{method-HourlyApi-v1_observations_hourly_station_id_date_time_start_get_with_http_info}{}}}
\subsection{Method \code{v1_observations_hourly_station_id_date_time_start_get_with_http_info()}}{
Return hourly direct and derived observations for a specific station and date and time.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HourlyApi$v1_observations_hourly_station_id_date_time_start_get_with_http_info(
  station_id,
  date_time_start,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{station_id}}{ID of station}

\item{\code{date_time_start}}{Date and time of data request start}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (InlineResponse2003) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HourlyApi-v1_observations_hourly_station_id_date_time_start_time_interval_get"></a>}}
\if{latex}{\out{\hypertarget{method-HourlyApi-v1_observations_hourly_station_id_date_time_start_time_interval_get}{}}}
\subsection{Method \code{v1_observations_hourly_station_id_date_time_start_time_interval_get()}}{
Return hourly direct and derived observations for a specific station over a time interval from a date and time.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HourlyApi$v1_observations_hourly_station_id_date_time_start_time_interval_get(
  station_id,
  date_time_start,
  time_interval,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{station_id}}{ID of station}

\item{\code{date_time_start}}{Date and time of data request start}

\item{\code{time_interval}}{ID of station}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
InlineResponse2004
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HourlyApi-v1_observations_hourly_station_id_date_time_start_time_interval_get_with_http_info"></a>}}
\if{latex}{\out{\hypertarget{method-HourlyApi-v1_observations_hourly_station_id_date_time_start_time_interval_get_with_http_info}{}}}
\subsection{Method \code{v1_observations_hourly_station_id_date_time_start_time_interval_get_with_http_info()}}{
Return hourly direct and derived observations for a specific station over a time interval from a date and time.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HourlyApi$v1_observations_hourly_station_id_date_time_start_time_interval_get_with_http_info(
  station_id,
  date_time_start,
  time_interval,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{station_id}}{ID of station}

\item{\code{date_time_start}}{Date and time of data request start}

\item{\code{time_interval}}{ID of station}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (InlineResponse2004) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HourlyApi-v1_observations_hourly_station_id_get"></a>}}
\if{latex}{\out{\hypertarget{method-HourlyApi-v1_observations_hourly_station_id_get}{}}}
\subsection{Method \code{v1_observations_hourly_station_id_get()}}{
Return hourly direct and derived observations for a specific station.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HourlyApi$v1_observations_hourly_station_id_get(
  station_id,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{station_id}}{ID of station}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
InlineResponse2003
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HourlyApi-v1_observations_hourly_station_id_get_with_http_info"></a>}}
\if{latex}{\out{\hypertarget{method-HourlyApi-v1_observations_hourly_station_id_get_with_http_info}{}}}
\subsection{Method \code{v1_observations_hourly_station_id_get_with_http_info()}}{
Return hourly direct and derived observations for a specific station.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HourlyApi$v1_observations_hourly_station_id_get_with_http_info(
  station_id,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{station_id}}{ID of station}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (InlineResponse2003) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HourlyApi-clone"></a>}}
\if{latex}{\out{\hypertarget{method-HourlyApi-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HourlyApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
