% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hueto_api.R
\docType{class}
\name{HuetoApi}
\alias{HuetoApi}
\title{Hueto operations}
\format{
An \code{R6Class} generator object
}
\description{
HuetoApi
}
\details{
AZMet API

REST API for the Arizona Meteorological Network (AZMet)

The version of the OpenAPI document: 1.0
Generated by: https://openapi-generator.tech
}
\section{Methods}{

\describe{
\strong{ v1_observations_hueto_get } \emph{ Accumulated heat units and evapotranspiration for all stations from January 1 current year. }
Returns accumulated heat units and evapotranspiration for all stations from the start of the current year.

\itemize{
\item \emph{ @returnType } \link{InlineResponse2005} \cr

\item On encountering errors, an error of subclass ApiException will be thrown.

\item status code : 200 | Accumulated heat units and evapotranspiration for all stations from the start of the current year.

\item return type : InlineResponse2005
\item response headers :

\tabular{ll}{
}
}

\strong{ v1_observations_hueto_station_id_date_time_start_get } \emph{ Accumulated heat units and evapotranspiration for one (or all) station(s) from the given start date and time to current date. }
Returns accumulated heat units and evapotranspiration for one (or all \"*\") station(s) from the given start date and time to current date.

\itemize{
\item \emph{ @param } station_id character
\item \emph{ @param } date_time_start character
\item \emph{ @returnType } \link{InlineResponse2006} \cr

\item On encountering errors, an error of subclass ApiException will be thrown.

\item status code : 200 | Auto generated using Swagger Inspector

\item return type : InlineResponse2006
\item response headers :

\tabular{ll}{
}
}

\strong{ v1_observations_hueto_station_id_date_time_start_time_interval_get } \emph{ Accumulated heat units and evapotranspiration for one (or all) station(s) from the given start date and time over a given time interval. }
Returns accumulated heat units and evapotranspiration for one (or all \"*\") station(s) from the given start date and timeover a given time interval.

\itemize{
\item \emph{ @param } station_id character
\item \emph{ @param } date_time_start character
\item \emph{ @param } time_interval character
\item \emph{ @returnType } \link{InlineResponse2006} \cr

\item On encountering errors, an error of subclass ApiException will be thrown.

\item status code : 200 | Auto generated using Swagger Inspector

\item return type : InlineResponse2006
\item response headers :

\tabular{ll}{
}
}

\strong{ v1_observations_hueto_station_id_get } \emph{ Accumulated heat units and evapotranspiration for one station from January 1 current year. }
Returns accumulated heat units and evapotranspiration for one station from the start of the current year.

\itemize{
\item \emph{ @param } station_id character
\item \emph{ @returnType } \link{InlineResponse2006} \cr

\item On encountering errors, an error of subclass ApiException will be thrown.

\item status code : 200 | Accumulated heat units and evapotranspiration for one station from the start of the current year.

\item return type : InlineResponse2006
\item response headers :

\tabular{ll}{
}
}

}
}

\examples{
\dontrun{
####################  v1_observations_hueto_get  ####################

library(azmet)

#Accumulated heat units and evapotranspiration for all stations from January 1 current year.
api_instance <- HuetoApi$new()

result <- tryCatch(
             
             # to save the result into a file, simply add the optional `data_file` parameter, e.g.
             # api_instance$v1_observations_hueto_get(data_file = "result.txt"),
             
             
             api_instance$v1_observations_hueto_get(),
             ApiException = function(ex) ex
          )
# In case of error, print the error object
if (!is.null(result$ApiException)) {
  print("Exception occurs when calling `v1_observations_hueto_get`:")
  dput(result$ApiException$toString())
  
  # error object
  dput(result$ApiException$error_object$toJSONString())
  
} else {
  # deserialized response object
  print("The response is ...")
  dput(result$toString())
}



####################  v1_observations_hueto_station_id_date_time_start_get  ####################

library(azmet)
var_station_id <- "station_id_example" # character | ID of station
var_date_time_start <- "date_time_start_example" # character | Date and time of data request start

#Accumulated heat units and evapotranspiration for one (or all) station(s) from the given start date and time to current date.
api_instance <- HuetoApi$new()

result <- tryCatch(
             
             # to save the result into a file, simply add the optional `data_file` parameter, e.g.
             # api_instance$v1_observations_hueto_station_id_date_time_start_get(var_station_id, var_date_time_start, data_file = "result.txt"),
             
             
             api_instance$v1_observations_hueto_station_id_date_time_start_get(var_station_id, var_date_time_start),
             ApiException = function(ex) ex
          )
# In case of error, print the error object
if (!is.null(result$ApiException)) {
  print("Exception occurs when calling `v1_observations_hueto_station_id_date_time_start_get`:")
  dput(result$ApiException$toString())
  
  # error object
  dput(result$ApiException$error_object$toJSONString())
  
} else {
  # deserialized response object
  print("The response is ...")
  dput(result$toString())
}



####################  v1_observations_hueto_station_id_date_time_start_time_interval_get  ####################

library(azmet)
var_station_id <- "station_id_example" # character | ID of station
var_date_time_start <- "date_time_start_example" # character | Date and time of data request start
var_time_interval <- "time_interval_example" # character | The requested time interval

#Accumulated heat units and evapotranspiration for one (or all) station(s) from the given start date and time over a given time interval.
api_instance <- HuetoApi$new()

result <- tryCatch(
             
             # to save the result into a file, simply add the optional `data_file` parameter, e.g.
             # api_instance$v1_observations_hueto_station_id_date_time_start_time_interval_get(var_station_id, var_date_time_start, var_time_interval, data_file = "result.txt"),
             
             
             api_instance$v1_observations_hueto_station_id_date_time_start_time_interval_get(var_station_id, var_date_time_start, var_time_interval),
             ApiException = function(ex) ex
          )
# In case of error, print the error object
if (!is.null(result$ApiException)) {
  print("Exception occurs when calling `v1_observations_hueto_station_id_date_time_start_time_interval_get`:")
  dput(result$ApiException$toString())
  
  # error object
  dput(result$ApiException$error_object$toJSONString())
  
} else {
  # deserialized response object
  print("The response is ...")
  dput(result$toString())
}



####################  v1_observations_hueto_station_id_get  ####################

library(azmet)
var_station_id <- "station_id_example" # character | ID of station

#Accumulated heat units and evapotranspiration for one station from January 1 current year.
api_instance <- HuetoApi$new()

result <- tryCatch(
             
             # to save the result into a file, simply add the optional `data_file` parameter, e.g.
             # api_instance$v1_observations_hueto_station_id_get(var_station_id, data_file = "result.txt"),
             
             
             api_instance$v1_observations_hueto_station_id_get(var_station_id),
             ApiException = function(ex) ex
          )
# In case of error, print the error object
if (!is.null(result$ApiException)) {
  print("Exception occurs when calling `v1_observations_hueto_station_id_get`:")
  dput(result$ApiException$toString())
  
  # error object
  dput(result$ApiException$error_object$toJSONString())
  
} else {
  # deserialized response object
  print("The response is ...")
  dput(result$toString())
}



}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{api_client}}{Handles the client-server communication.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-HuetoApi-new}{\code{HuetoApi$new()}}
\item \href{#method-HuetoApi-v1_observations_hueto_get}{\code{HuetoApi$v1_observations_hueto_get()}}
\item \href{#method-HuetoApi-v1_observations_hueto_get_with_http_info}{\code{HuetoApi$v1_observations_hueto_get_with_http_info()}}
\item \href{#method-HuetoApi-v1_observations_hueto_station_id_date_time_start_get}{\code{HuetoApi$v1_observations_hueto_station_id_date_time_start_get()}}
\item \href{#method-HuetoApi-v1_observations_hueto_station_id_date_time_start_get_with_http_info}{\code{HuetoApi$v1_observations_hueto_station_id_date_time_start_get_with_http_info()}}
\item \href{#method-HuetoApi-v1_observations_hueto_station_id_date_time_start_time_interval_get}{\code{HuetoApi$v1_observations_hueto_station_id_date_time_start_time_interval_get()}}
\item \href{#method-HuetoApi-v1_observations_hueto_station_id_date_time_start_time_interval_get_with_http_info}{\code{HuetoApi$v1_observations_hueto_station_id_date_time_start_time_interval_get_with_http_info()}}
\item \href{#method-HuetoApi-v1_observations_hueto_station_id_get}{\code{HuetoApi$v1_observations_hueto_station_id_get()}}
\item \href{#method-HuetoApi-v1_observations_hueto_station_id_get_with_http_info}{\code{HuetoApi$v1_observations_hueto_station_id_get_with_http_info()}}
\item \href{#method-HuetoApi-clone}{\code{HuetoApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HuetoApi-new"></a>}}
\if{latex}{\out{\hypertarget{method-HuetoApi-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new HuetoApi.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HuetoApi$new(api_client)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{api_client}}{An instance of API client.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HuetoApi-v1_observations_hueto_get"></a>}}
\if{latex}{\out{\hypertarget{method-HuetoApi-v1_observations_hueto_get}{}}}
\subsection{Method \code{v1_observations_hueto_get()}}{
Accumulated heat units and evapotranspiration for all stations from January 1 current year.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HuetoApi$v1_observations_hueto_get(data_file = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
InlineResponse2005
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HuetoApi-v1_observations_hueto_get_with_http_info"></a>}}
\if{latex}{\out{\hypertarget{method-HuetoApi-v1_observations_hueto_get_with_http_info}{}}}
\subsection{Method \code{v1_observations_hueto_get_with_http_info()}}{
Accumulated heat units and evapotranspiration for all stations from January 1 current year.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HuetoApi$v1_observations_hueto_get_with_http_info(data_file = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (InlineResponse2005) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HuetoApi-v1_observations_hueto_station_id_date_time_start_get"></a>}}
\if{latex}{\out{\hypertarget{method-HuetoApi-v1_observations_hueto_station_id_date_time_start_get}{}}}
\subsection{Method \code{v1_observations_hueto_station_id_date_time_start_get()}}{
Accumulated heat units and evapotranspiration for one (or all) station(s) from the given start date and time to current date.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HuetoApi$v1_observations_hueto_station_id_date_time_start_get(
  station_id,
  date_time_start,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{station_id}}{ID of station}

\item{\code{date_time_start}}{Date and time of data request start}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
InlineResponse2006
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HuetoApi-v1_observations_hueto_station_id_date_time_start_get_with_http_info"></a>}}
\if{latex}{\out{\hypertarget{method-HuetoApi-v1_observations_hueto_station_id_date_time_start_get_with_http_info}{}}}
\subsection{Method \code{v1_observations_hueto_station_id_date_time_start_get_with_http_info()}}{
Accumulated heat units and evapotranspiration for one (or all) station(s) from the given start date and time to current date.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HuetoApi$v1_observations_hueto_station_id_date_time_start_get_with_http_info(
  station_id,
  date_time_start,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{station_id}}{ID of station}

\item{\code{date_time_start}}{Date and time of data request start}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (InlineResponse2006) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HuetoApi-v1_observations_hueto_station_id_date_time_start_time_interval_get"></a>}}
\if{latex}{\out{\hypertarget{method-HuetoApi-v1_observations_hueto_station_id_date_time_start_time_interval_get}{}}}
\subsection{Method \code{v1_observations_hueto_station_id_date_time_start_time_interval_get()}}{
Accumulated heat units and evapotranspiration for one (or all) station(s) from the given start date and time over a given time interval.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HuetoApi$v1_observations_hueto_station_id_date_time_start_time_interval_get(
  station_id,
  date_time_start,
  time_interval,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{station_id}}{ID of station}

\item{\code{date_time_start}}{Date and time of data request start}

\item{\code{time_interval}}{The requested time interval}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
InlineResponse2006
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HuetoApi-v1_observations_hueto_station_id_date_time_start_time_interval_get_with_http_info"></a>}}
\if{latex}{\out{\hypertarget{method-HuetoApi-v1_observations_hueto_station_id_date_time_start_time_interval_get_with_http_info}{}}}
\subsection{Method \code{v1_observations_hueto_station_id_date_time_start_time_interval_get_with_http_info()}}{
Accumulated heat units and evapotranspiration for one (or all) station(s) from the given start date and time over a given time interval.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HuetoApi$v1_observations_hueto_station_id_date_time_start_time_interval_get_with_http_info(
  station_id,
  date_time_start,
  time_interval,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{station_id}}{ID of station}

\item{\code{date_time_start}}{Date and time of data request start}

\item{\code{time_interval}}{The requested time interval}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (InlineResponse2006) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HuetoApi-v1_observations_hueto_station_id_get"></a>}}
\if{latex}{\out{\hypertarget{method-HuetoApi-v1_observations_hueto_station_id_get}{}}}
\subsection{Method \code{v1_observations_hueto_station_id_get()}}{
Accumulated heat units and evapotranspiration for one station from January 1 current year.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HuetoApi$v1_observations_hueto_station_id_get(
  station_id,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{station_id}}{ID of station}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
InlineResponse2006
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HuetoApi-v1_observations_hueto_station_id_get_with_http_info"></a>}}
\if{latex}{\out{\hypertarget{method-HuetoApi-v1_observations_hueto_station_id_get_with_http_info}{}}}
\subsection{Method \code{v1_observations_hueto_station_id_get_with_http_info()}}{
Accumulated heat units and evapotranspiration for one station from January 1 current year.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HuetoApi$v1_observations_hueto_station_id_get_with_http_info(
  station_id,
  data_file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{station_id}}{ID of station}

\item{\code{data_file}}{(optional) name of the data file to save the result}

\item{\code{...}}{Other optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
API response (InlineResponse2006) with additional information such as HTTP status code, headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HuetoApi-clone"></a>}}
\if{latex}{\out{\hypertarget{method-HuetoApi-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HuetoApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
