% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inline_response200.R
\docType{class}
\name{InlineResponse200}
\alias{InlineResponse200}
\title{InlineResponse200}
\format{
An \code{R6Class} generator object
}
\description{
InlineResponse200 Class

InlineResponse200 Class
}
\details{
Create a new InlineResponse200
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{list(\link{InlineResponse200DataInner}) [optional]}

\item{\code{errors}}{list(character) [optional]}

\item{\code{i}}{character [optional]}

\item{\code{l}}{character [optional]}

\item{\code{s}}{character [optional]}

\item{\code{t}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-InlineResponse200-new}{\code{InlineResponse200$new()}}
\item \href{#method-InlineResponse200-toJSON}{\code{InlineResponse200$toJSON()}}
\item \href{#method-InlineResponse200-fromJSON}{\code{InlineResponse200$fromJSON()}}
\item \href{#method-InlineResponse200-toJSONString}{\code{InlineResponse200$toJSONString()}}
\item \href{#method-InlineResponse200-fromJSONString}{\code{InlineResponse200$fromJSONString()}}
\item \href{#method-InlineResponse200-validateJSON}{\code{InlineResponse200$validateJSON()}}
\item \href{#method-InlineResponse200-toString}{\code{InlineResponse200$toString()}}
\item \href{#method-InlineResponse200-isValid}{\code{InlineResponse200$isValid()}}
\item \href{#method-InlineResponse200-getInvalidFields}{\code{InlineResponse200$getInvalidFields()}}
\item \href{#method-InlineResponse200-print}{\code{InlineResponse200$print()}}
\item \href{#method-InlineResponse200-clone}{\code{InlineResponse200$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InlineResponse200-new"></a>}}
\if{latex}{\out{\hypertarget{method-InlineResponse200-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new InlineResponse200 class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse200$new(
  data = NULL,
  errors = NULL,
  i = NULL,
  l = NULL,
  s = NULL,
  t = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{data}

\item{\code{errors}}{errors}

\item{\code{i}}{i}

\item{\code{l}}{l}

\item{\code{s}}{s}

\item{\code{t}}{t}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InlineResponse200-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-InlineResponse200-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse200$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
InlineResponse200 in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InlineResponse200-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-InlineResponse200-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of InlineResponse200
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse200$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of InlineResponse200
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InlineResponse200-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-InlineResponse200-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse200$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
InlineResponse200 in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InlineResponse200-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-InlineResponse200-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of InlineResponse200
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse200$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of InlineResponse200
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InlineResponse200-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-InlineResponse200-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to InlineResponse200 and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse200$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InlineResponse200-toString"></a>}}
\if{latex}{\out{\hypertarget{method-InlineResponse200-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse200$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of InlineResponse200
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InlineResponse200-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-InlineResponse200-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse200$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InlineResponse200-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-InlineResponse200-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse200$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InlineResponse200-print"></a>}}
\if{latex}{\out{\hypertarget{method-InlineResponse200-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse200$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InlineResponse200-clone"></a>}}
\if{latex}{\out{\hypertarget{method-InlineResponse200-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse200$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
