% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inline_response2001.R
\docType{class}
\name{InlineResponse2001}
\alias{InlineResponse2001}
\title{InlineResponse2001}
\format{
An \code{R6Class} generator object
}
\description{
InlineResponse2001 Class

InlineResponse2001 Class
}
\details{
Create a new InlineResponse2001
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{a}}{character [optional]}

\item{\code{s}}{character [optional]}

\item{\code{data}}{list(\link{InlineResponse2001DataInner}) [optional]}

\item{\code{t}}{character [optional]}

\item{\code{i}}{character [optional]}

\item{\code{l}}{character [optional]}

\item{\code{errors}}{list(character) [optional]}

\item{\code{n}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-InlineResponse2001-new}{\code{InlineResponse2001$new()}}
\item \href{#method-InlineResponse2001-toJSON}{\code{InlineResponse2001$toJSON()}}
\item \href{#method-InlineResponse2001-fromJSON}{\code{InlineResponse2001$fromJSON()}}
\item \href{#method-InlineResponse2001-toJSONString}{\code{InlineResponse2001$toJSONString()}}
\item \href{#method-InlineResponse2001-fromJSONString}{\code{InlineResponse2001$fromJSONString()}}
\item \href{#method-InlineResponse2001-validateJSON}{\code{InlineResponse2001$validateJSON()}}
\item \href{#method-InlineResponse2001-toString}{\code{InlineResponse2001$toString()}}
\item \href{#method-InlineResponse2001-isValid}{\code{InlineResponse2001$isValid()}}
\item \href{#method-InlineResponse2001-getInvalidFields}{\code{InlineResponse2001$getInvalidFields()}}
\item \href{#method-InlineResponse2001-print}{\code{InlineResponse2001$print()}}
\item \href{#method-InlineResponse2001-clone}{\code{InlineResponse2001$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InlineResponse2001-new"></a>}}
\if{latex}{\out{\hypertarget{method-InlineResponse2001-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new InlineResponse2001 class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse2001$new(
  a = NULL,
  s = NULL,
  data = NULL,
  t = NULL,
  i = NULL,
  l = NULL,
  errors = NULL,
  n = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{a}}{a}

\item{\code{s}}{s}

\item{\code{data}}{data}

\item{\code{t}}{t}

\item{\code{i}}{i}

\item{\code{l}}{l}

\item{\code{errors}}{errors}

\item{\code{n}}{n}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InlineResponse2001-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-InlineResponse2001-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse2001$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
InlineResponse2001 in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InlineResponse2001-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-InlineResponse2001-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of InlineResponse2001
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse2001$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of InlineResponse2001
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InlineResponse2001-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-InlineResponse2001-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse2001$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
InlineResponse2001 in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InlineResponse2001-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-InlineResponse2001-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of InlineResponse2001
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse2001$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of InlineResponse2001
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InlineResponse2001-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-InlineResponse2001-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to InlineResponse2001 and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse2001$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InlineResponse2001-toString"></a>}}
\if{latex}{\out{\hypertarget{method-InlineResponse2001-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse2001$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of InlineResponse2001
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InlineResponse2001-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-InlineResponse2001-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse2001$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InlineResponse2001-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-InlineResponse2001-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse2001$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InlineResponse2001-print"></a>}}
\if{latex}{\out{\hypertarget{method-InlineResponse2001-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse2001$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InlineResponse2001-clone"></a>}}
\if{latex}{\out{\hypertarget{method-InlineResponse2001-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse2001$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
