% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inline_response2001_data_inner.R
\docType{class}
\name{InlineResponse2001DataInner}
\alias{InlineResponse2001DataInner}
\title{InlineResponse2001DataInner}
\format{
An \code{R6Class} generator object
}
\description{
InlineResponse2001DataInner Class

InlineResponse2001DataInner Class
}
\details{
Create a new InlineResponse2001DataInner
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{temp_air_meanC}}{character [optional]}

\item{\code{date_year}}{character [optional]}

\item{\code{temp_air_minC}}{character [optional]}

\item{\code{heat_units_3413C}}{character [optional]}

\item{\code{temp_soil_10cm_maxF}}{character [optional]}

\item{\code{meta_version}}{integer [optional]}

\item{\code{temp_air_minF}}{character [optional]}

\item{\code{temp_soil_10cm_maxC}}{character [optional]}

\item{\code{relative_humidity_max}}{character [optional]}

\item{\code{wind_vector_dir}}{character [optional]}

\item{\code{datetime}}{character [optional]}

\item{\code{heat_units_9455F}}{character [optional]}

\item{\code{dwpt_meanF}}{character [optional]}

\item{\code{meta_station_id}}{character [optional]}

\item{\code{temp_soil_50cm_maxC}}{character [optional]}

\item{\code{wind_spd_mean_mps}}{character [optional]}

\item{\code{temp_soil_50cm_maxF}}{character [optional]}

\item{\code{heat_units_55F}}{character [optional]}

\item{\code{wind_spd_mean_mph}}{character [optional]}

\item{\code{sol_rad_total}}{character [optional]}

\item{\code{chill_hours_68F}}{character [optional]}

\item{\code{precip_total_in}}{character [optional]}

\item{\code{temp_soil_50cm_meanF}}{character [optional]}

\item{\code{chill_hours_20C}}{character [optional]}

\item{\code{meta_needs_review}}{integer [optional]}

\item{\code{sol_rad_total_ly}}{character [optional]}

\item{\code{dwpt_mean}}{character [optional]}

\item{\code{eto_azmet}}{character [optional]}

\item{\code{chill_hours_32F}}{character [optional]}

\item{\code{temp_air_maxC}}{character [optional]}

\item{\code{temp_soil_50cm_meanC}}{character [optional]}

\item{\code{temp_soil_10cm_meanF}}{character [optional]}

\item{\code{wind_vector_dir_stand_dev}}{character [optional]}

\item{\code{heatstress_cotton_meanF}}{character [optional]}

\item{\code{wind_spd_max_mps}}{character [optional]}

\item{\code{relative_humidity_min}}{character [optional]}

\item{\code{heatstress_cotton_meanC}}{character [optional]}

\item{\code{meta_bat_volt_min}}{character [optional]}

\item{\code{meta_station_name}}{character [optional]}

\item{\code{vp_actual_mean}}{character [optional]}

\item{\code{temp_soil_50cm_minF}}{character [optional]}

\item{\code{heat_units_10C}}{character [optional]}

\item{\code{relative_humidity_mean}}{character [optional]}

\item{\code{meta_bat_volt_max}}{character [optional]}

\item{\code{wind_spd_max_mph}}{character [optional]}

\item{\code{heat_units_50F}}{character [optional]}

\item{\code{chill_hours_45F}}{character [optional]}

\item{\code{eto_pen_mon}}{character [optional]}

\item{\code{temp_soil_10cm_meanC}}{character [optional]}

\item{\code{temp_soil_50cm_minC}}{character [optional]}

\item{\code{temp_air_maxF}}{character [optional]}

\item{\code{temp_soil_10cm_minF}}{character [optional]}

\item{\code{temp_soil_10cm_minC}}{character [optional]}

\item{\code{chill_hours_0C}}{character [optional]}

\item{\code{date_doy}}{character [optional]}

\item{\code{vp_deficit_mean}}{character [optional]}

\item{\code{heat_units_45F}}{character [optional]}

\item{\code{eto_azmet_in}}{character [optional]}

\item{\code{heat_units_13C}}{character [optional]}

\item{\code{heat_units_7C}}{character [optional]}

\item{\code{vp_actual_min}}{character [optional]}

\item{\code{meta_bat_volt_mean}}{character [optional]}

\item{\code{wind_vector_magnitude_mph}}{character [optional]}

\item{\code{precip_total_mm}}{character [optional]}

\item{\code{vp_actual_max}}{character [optional]}

\item{\code{wind_vector_magnitude}}{character [optional]}

\item{\code{eto_pen_mon_in}}{character [optional]}

\item{\code{chill_hours_7C}}{character [optional]}

\item{\code{temp_air_meanF}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-InlineResponse2001DataInner-new}{\code{InlineResponse2001DataInner$new()}}
\item \href{#method-InlineResponse2001DataInner-toJSON}{\code{InlineResponse2001DataInner$toJSON()}}
\item \href{#method-InlineResponse2001DataInner-fromJSON}{\code{InlineResponse2001DataInner$fromJSON()}}
\item \href{#method-InlineResponse2001DataInner-toJSONString}{\code{InlineResponse2001DataInner$toJSONString()}}
\item \href{#method-InlineResponse2001DataInner-fromJSONString}{\code{InlineResponse2001DataInner$fromJSONString()}}
\item \href{#method-InlineResponse2001DataInner-validateJSON}{\code{InlineResponse2001DataInner$validateJSON()}}
\item \href{#method-InlineResponse2001DataInner-toString}{\code{InlineResponse2001DataInner$toString()}}
\item \href{#method-InlineResponse2001DataInner-isValid}{\code{InlineResponse2001DataInner$isValid()}}
\item \href{#method-InlineResponse2001DataInner-getInvalidFields}{\code{InlineResponse2001DataInner$getInvalidFields()}}
\item \href{#method-InlineResponse2001DataInner-print}{\code{InlineResponse2001DataInner$print()}}
\item \href{#method-InlineResponse2001DataInner-clone}{\code{InlineResponse2001DataInner$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InlineResponse2001DataInner-new"></a>}}
\if{latex}{\out{\hypertarget{method-InlineResponse2001DataInner-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new InlineResponse2001DataInner class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse2001DataInner$new(
  temp_air_meanC = NULL,
  date_year = NULL,
  temp_air_minC = NULL,
  heat_units_3413C = NULL,
  temp_soil_10cm_maxF = NULL,
  meta_version = NULL,
  temp_air_minF = NULL,
  temp_soil_10cm_maxC = NULL,
  relative_humidity_max = NULL,
  wind_vector_dir = NULL,
  datetime = NULL,
  heat_units_9455F = NULL,
  dwpt_meanF = NULL,
  meta_station_id = NULL,
  temp_soil_50cm_maxC = NULL,
  wind_spd_mean_mps = NULL,
  temp_soil_50cm_maxF = NULL,
  heat_units_55F = NULL,
  wind_spd_mean_mph = NULL,
  sol_rad_total = NULL,
  chill_hours_68F = NULL,
  precip_total_in = NULL,
  temp_soil_50cm_meanF = NULL,
  chill_hours_20C = NULL,
  meta_needs_review = NULL,
  sol_rad_total_ly = NULL,
  dwpt_mean = NULL,
  eto_azmet = NULL,
  chill_hours_32F = NULL,
  temp_air_maxC = NULL,
  temp_soil_50cm_meanC = NULL,
  temp_soil_10cm_meanF = NULL,
  wind_vector_dir_stand_dev = NULL,
  heatstress_cotton_meanF = NULL,
  wind_spd_max_mps = NULL,
  relative_humidity_min = NULL,
  heatstress_cotton_meanC = NULL,
  meta_bat_volt_min = NULL,
  meta_station_name = NULL,
  vp_actual_mean = NULL,
  temp_soil_50cm_minF = NULL,
  heat_units_10C = NULL,
  relative_humidity_mean = NULL,
  meta_bat_volt_max = NULL,
  wind_spd_max_mph = NULL,
  heat_units_50F = NULL,
  chill_hours_45F = NULL,
  eto_pen_mon = NULL,
  temp_soil_10cm_meanC = NULL,
  temp_soil_50cm_minC = NULL,
  temp_air_maxF = NULL,
  temp_soil_10cm_minF = NULL,
  temp_soil_10cm_minC = NULL,
  chill_hours_0C = NULL,
  date_doy = NULL,
  vp_deficit_mean = NULL,
  heat_units_45F = NULL,
  eto_azmet_in = NULL,
  heat_units_13C = NULL,
  heat_units_7C = NULL,
  vp_actual_min = NULL,
  meta_bat_volt_mean = NULL,
  wind_vector_magnitude_mph = NULL,
  precip_total_mm = NULL,
  vp_actual_max = NULL,
  wind_vector_magnitude = NULL,
  eto_pen_mon_in = NULL,
  chill_hours_7C = NULL,
  temp_air_meanF = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{temp_air_meanC}}{temp_air_meanC}

\item{\code{date_year}}{date_year}

\item{\code{temp_air_minC}}{temp_air_minC}

\item{\code{heat_units_3413C}}{heat_units_3413C}

\item{\code{temp_soil_10cm_maxF}}{temp_soil_10cm_maxF}

\item{\code{meta_version}}{meta_version}

\item{\code{temp_air_minF}}{temp_air_minF}

\item{\code{temp_soil_10cm_maxC}}{temp_soil_10cm_maxC}

\item{\code{relative_humidity_max}}{relative_humidity_max}

\item{\code{wind_vector_dir}}{wind_vector_dir}

\item{\code{datetime}}{datetime}

\item{\code{heat_units_9455F}}{heat_units_9455F}

\item{\code{dwpt_meanF}}{dwpt_meanF}

\item{\code{meta_station_id}}{meta_station_id}

\item{\code{temp_soil_50cm_maxC}}{temp_soil_50cm_maxC}

\item{\code{wind_spd_mean_mps}}{wind_spd_mean_mps}

\item{\code{temp_soil_50cm_maxF}}{temp_soil_50cm_maxF}

\item{\code{heat_units_55F}}{heat_units_55F}

\item{\code{wind_spd_mean_mph}}{wind_spd_mean_mph}

\item{\code{sol_rad_total}}{sol_rad_total}

\item{\code{chill_hours_68F}}{chill_hours_68F}

\item{\code{precip_total_in}}{precip_total_in}

\item{\code{temp_soil_50cm_meanF}}{temp_soil_50cm_meanF}

\item{\code{chill_hours_20C}}{chill_hours_20C}

\item{\code{meta_needs_review}}{meta_needs_review}

\item{\code{sol_rad_total_ly}}{sol_rad_total_ly}

\item{\code{dwpt_mean}}{dwpt_mean}

\item{\code{eto_azmet}}{eto_azmet}

\item{\code{chill_hours_32F}}{chill_hours_32F}

\item{\code{temp_air_maxC}}{temp_air_maxC}

\item{\code{temp_soil_50cm_meanC}}{temp_soil_50cm_meanC}

\item{\code{temp_soil_10cm_meanF}}{temp_soil_10cm_meanF}

\item{\code{wind_vector_dir_stand_dev}}{wind_vector_dir_stand_dev}

\item{\code{heatstress_cotton_meanF}}{heatstress_cotton_meanF}

\item{\code{wind_spd_max_mps}}{wind_spd_max_mps}

\item{\code{relative_humidity_min}}{relative_humidity_min}

\item{\code{heatstress_cotton_meanC}}{heatstress_cotton_meanC}

\item{\code{meta_bat_volt_min}}{meta_bat_volt_min}

\item{\code{meta_station_name}}{meta_station_name}

\item{\code{vp_actual_mean}}{vp_actual_mean}

\item{\code{temp_soil_50cm_minF}}{temp_soil_50cm_minF}

\item{\code{heat_units_10C}}{heat_units_10C}

\item{\code{relative_humidity_mean}}{relative_humidity_mean}

\item{\code{meta_bat_volt_max}}{meta_bat_volt_max}

\item{\code{wind_spd_max_mph}}{wind_spd_max_mph}

\item{\code{heat_units_50F}}{heat_units_50F}

\item{\code{chill_hours_45F}}{chill_hours_45F}

\item{\code{eto_pen_mon}}{eto_pen_mon}

\item{\code{temp_soil_10cm_meanC}}{temp_soil_10cm_meanC}

\item{\code{temp_soil_50cm_minC}}{temp_soil_50cm_minC}

\item{\code{temp_air_maxF}}{temp_air_maxF}

\item{\code{temp_soil_10cm_minF}}{temp_soil_10cm_minF}

\item{\code{temp_soil_10cm_minC}}{temp_soil_10cm_minC}

\item{\code{chill_hours_0C}}{chill_hours_0C}

\item{\code{date_doy}}{date_doy}

\item{\code{vp_deficit_mean}}{vp_deficit_mean}

\item{\code{heat_units_45F}}{heat_units_45F}

\item{\code{eto_azmet_in}}{eto_azmet_in}

\item{\code{heat_units_13C}}{heat_units_13C}

\item{\code{heat_units_7C}}{heat_units_7C}

\item{\code{vp_actual_min}}{vp_actual_min}

\item{\code{meta_bat_volt_mean}}{meta_bat_volt_mean}

\item{\code{wind_vector_magnitude_mph}}{wind_vector_magnitude_mph}

\item{\code{precip_total_mm}}{precip_total_mm}

\item{\code{vp_actual_max}}{vp_actual_max}

\item{\code{wind_vector_magnitude}}{wind_vector_magnitude}

\item{\code{eto_pen_mon_in}}{eto_pen_mon_in}

\item{\code{chill_hours_7C}}{chill_hours_7C}

\item{\code{temp_air_meanF}}{temp_air_meanF}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InlineResponse2001DataInner-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-InlineResponse2001DataInner-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse2001DataInner$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
InlineResponse2001DataInner in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InlineResponse2001DataInner-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-InlineResponse2001DataInner-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of InlineResponse2001DataInner
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse2001DataInner$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of InlineResponse2001DataInner
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InlineResponse2001DataInner-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-InlineResponse2001DataInner-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse2001DataInner$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
InlineResponse2001DataInner in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InlineResponse2001DataInner-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-InlineResponse2001DataInner-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of InlineResponse2001DataInner
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse2001DataInner$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of InlineResponse2001DataInner
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InlineResponse2001DataInner-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-InlineResponse2001DataInner-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to InlineResponse2001DataInner and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse2001DataInner$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InlineResponse2001DataInner-toString"></a>}}
\if{latex}{\out{\hypertarget{method-InlineResponse2001DataInner-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse2001DataInner$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of InlineResponse2001DataInner
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InlineResponse2001DataInner-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-InlineResponse2001DataInner-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse2001DataInner$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InlineResponse2001DataInner-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-InlineResponse2001DataInner-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse2001DataInner$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InlineResponse2001DataInner-print"></a>}}
\if{latex}{\out{\hypertarget{method-InlineResponse2001DataInner-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse2001DataInner$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InlineResponse2001DataInner-clone"></a>}}
\if{latex}{\out{\hypertarget{method-InlineResponse2001DataInner-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse2001DataInner$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
