% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inline_response2002.R
\docType{class}
\name{InlineResponse2002}
\alias{InlineResponse2002}
\title{InlineResponse2002}
\format{
An \code{R6Class} generator object
}
\description{
InlineResponse2002 Class

InlineResponse2002 Class
}
\details{
Create a new InlineResponse2002
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{list(\link{InlineResponse2002DataInner}) [optional]}

\item{\code{errors}}{list(character) [optional]}

\item{\code{i}}{character [optional]}

\item{\code{l}}{character [optional]}

\item{\code{s}}{character [optional]}

\item{\code{t}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-InlineResponse2002-new}{\code{InlineResponse2002$new()}}
\item \href{#method-InlineResponse2002-toJSON}{\code{InlineResponse2002$toJSON()}}
\item \href{#method-InlineResponse2002-fromJSON}{\code{InlineResponse2002$fromJSON()}}
\item \href{#method-InlineResponse2002-toJSONString}{\code{InlineResponse2002$toJSONString()}}
\item \href{#method-InlineResponse2002-fromJSONString}{\code{InlineResponse2002$fromJSONString()}}
\item \href{#method-InlineResponse2002-validateJSON}{\code{InlineResponse2002$validateJSON()}}
\item \href{#method-InlineResponse2002-toString}{\code{InlineResponse2002$toString()}}
\item \href{#method-InlineResponse2002-isValid}{\code{InlineResponse2002$isValid()}}
\item \href{#method-InlineResponse2002-getInvalidFields}{\code{InlineResponse2002$getInvalidFields()}}
\item \href{#method-InlineResponse2002-print}{\code{InlineResponse2002$print()}}
\item \href{#method-InlineResponse2002-clone}{\code{InlineResponse2002$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InlineResponse2002-new"></a>}}
\if{latex}{\out{\hypertarget{method-InlineResponse2002-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new InlineResponse2002 class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse2002$new(
  data = NULL,
  errors = NULL,
  i = NULL,
  l = NULL,
  s = NULL,
  t = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{data}

\item{\code{errors}}{errors}

\item{\code{i}}{i}

\item{\code{l}}{l}

\item{\code{s}}{s}

\item{\code{t}}{t}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InlineResponse2002-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-InlineResponse2002-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse2002$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
InlineResponse2002 in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InlineResponse2002-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-InlineResponse2002-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of InlineResponse2002
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse2002$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of InlineResponse2002
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InlineResponse2002-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-InlineResponse2002-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse2002$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
InlineResponse2002 in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InlineResponse2002-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-InlineResponse2002-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of InlineResponse2002
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse2002$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of InlineResponse2002
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InlineResponse2002-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-InlineResponse2002-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to InlineResponse2002 and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse2002$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InlineResponse2002-toString"></a>}}
\if{latex}{\out{\hypertarget{method-InlineResponse2002-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse2002$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of InlineResponse2002
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InlineResponse2002-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-InlineResponse2002-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse2002$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InlineResponse2002-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-InlineResponse2002-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse2002$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InlineResponse2002-print"></a>}}
\if{latex}{\out{\hypertarget{method-InlineResponse2002-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse2002$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InlineResponse2002-clone"></a>}}
\if{latex}{\out{\hypertarget{method-InlineResponse2002-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse2002$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
