% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inline_response2004.R
\docType{class}
\name{InlineResponse2004}
\alias{InlineResponse2004}
\title{InlineResponse2004}
\format{
An \code{R6Class} generator object
}
\description{
InlineResponse2004 Class

InlineResponse2004 Class
}
\details{
Create a new InlineResponse2004
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{a}}{character [optional]}

\item{\code{s}}{character [optional]}

\item{\code{data}}{list(\link{InlineResponse2003DataInner}) [optional]}

\item{\code{t}}{character [optional]}

\item{\code{i}}{character [optional]}

\item{\code{l}}{character [optional]}

\item{\code{errors}}{list(character) [optional]}

\item{\code{n}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-InlineResponse2004-new}{\code{InlineResponse2004$new()}}
\item \href{#method-InlineResponse2004-toJSON}{\code{InlineResponse2004$toJSON()}}
\item \href{#method-InlineResponse2004-fromJSON}{\code{InlineResponse2004$fromJSON()}}
\item \href{#method-InlineResponse2004-toJSONString}{\code{InlineResponse2004$toJSONString()}}
\item \href{#method-InlineResponse2004-fromJSONString}{\code{InlineResponse2004$fromJSONString()}}
\item \href{#method-InlineResponse2004-validateJSON}{\code{InlineResponse2004$validateJSON()}}
\item \href{#method-InlineResponse2004-toString}{\code{InlineResponse2004$toString()}}
\item \href{#method-InlineResponse2004-isValid}{\code{InlineResponse2004$isValid()}}
\item \href{#method-InlineResponse2004-getInvalidFields}{\code{InlineResponse2004$getInvalidFields()}}
\item \href{#method-InlineResponse2004-print}{\code{InlineResponse2004$print()}}
\item \href{#method-InlineResponse2004-clone}{\code{InlineResponse2004$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InlineResponse2004-new"></a>}}
\if{latex}{\out{\hypertarget{method-InlineResponse2004-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new InlineResponse2004 class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse2004$new(
  a = NULL,
  s = NULL,
  data = NULL,
  t = NULL,
  i = NULL,
  l = NULL,
  errors = NULL,
  n = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{a}}{a}

\item{\code{s}}{s}

\item{\code{data}}{data}

\item{\code{t}}{t}

\item{\code{i}}{i}

\item{\code{l}}{l}

\item{\code{errors}}{errors}

\item{\code{n}}{n}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InlineResponse2004-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-InlineResponse2004-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse2004$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
InlineResponse2004 in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InlineResponse2004-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-InlineResponse2004-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of InlineResponse2004
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse2004$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of InlineResponse2004
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InlineResponse2004-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-InlineResponse2004-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse2004$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
InlineResponse2004 in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InlineResponse2004-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-InlineResponse2004-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of InlineResponse2004
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse2004$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of InlineResponse2004
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InlineResponse2004-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-InlineResponse2004-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to InlineResponse2004 and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse2004$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InlineResponse2004-toString"></a>}}
\if{latex}{\out{\hypertarget{method-InlineResponse2004-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse2004$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of InlineResponse2004
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InlineResponse2004-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-InlineResponse2004-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse2004$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InlineResponse2004-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-InlineResponse2004-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse2004$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InlineResponse2004-print"></a>}}
\if{latex}{\out{\hypertarget{method-InlineResponse2004-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse2004$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InlineResponse2004-clone"></a>}}
\if{latex}{\out{\hypertarget{method-InlineResponse2004-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse2004$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
