% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inline_response2005.R
\docType{class}
\name{InlineResponse2005}
\alias{InlineResponse2005}
\title{InlineResponse2005}
\format{
An \code{R6Class} generator object
}
\description{
InlineResponse2005 Class

InlineResponse2005 Class
}
\details{
Create a new InlineResponse2005
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{list(\link{InlineResponse2005DataInner}) [optional]}

\item{\code{errors}}{list(character) [optional]}

\item{\code{i}}{character [optional]}

\item{\code{l}}{character [optional]}

\item{\code{s}}{character [optional]}

\item{\code{t}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-InlineResponse2005-new}{\code{InlineResponse2005$new()}}
\item \href{#method-InlineResponse2005-toJSON}{\code{InlineResponse2005$toJSON()}}
\item \href{#method-InlineResponse2005-fromJSON}{\code{InlineResponse2005$fromJSON()}}
\item \href{#method-InlineResponse2005-toJSONString}{\code{InlineResponse2005$toJSONString()}}
\item \href{#method-InlineResponse2005-fromJSONString}{\code{InlineResponse2005$fromJSONString()}}
\item \href{#method-InlineResponse2005-validateJSON}{\code{InlineResponse2005$validateJSON()}}
\item \href{#method-InlineResponse2005-toString}{\code{InlineResponse2005$toString()}}
\item \href{#method-InlineResponse2005-isValid}{\code{InlineResponse2005$isValid()}}
\item \href{#method-InlineResponse2005-getInvalidFields}{\code{InlineResponse2005$getInvalidFields()}}
\item \href{#method-InlineResponse2005-print}{\code{InlineResponse2005$print()}}
\item \href{#method-InlineResponse2005-clone}{\code{InlineResponse2005$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InlineResponse2005-new"></a>}}
\if{latex}{\out{\hypertarget{method-InlineResponse2005-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new InlineResponse2005 class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse2005$new(
  data = NULL,
  errors = NULL,
  i = NULL,
  l = NULL,
  s = NULL,
  t = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{data}

\item{\code{errors}}{errors}

\item{\code{i}}{i}

\item{\code{l}}{l}

\item{\code{s}}{s}

\item{\code{t}}{t}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InlineResponse2005-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-InlineResponse2005-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse2005$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
InlineResponse2005 in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InlineResponse2005-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-InlineResponse2005-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of InlineResponse2005
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse2005$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of InlineResponse2005
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InlineResponse2005-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-InlineResponse2005-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse2005$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
InlineResponse2005 in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InlineResponse2005-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-InlineResponse2005-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of InlineResponse2005
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse2005$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of InlineResponse2005
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InlineResponse2005-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-InlineResponse2005-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to InlineResponse2005 and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse2005$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InlineResponse2005-toString"></a>}}
\if{latex}{\out{\hypertarget{method-InlineResponse2005-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse2005$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of InlineResponse2005
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InlineResponse2005-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-InlineResponse2005-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse2005$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InlineResponse2005-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-InlineResponse2005-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse2005$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InlineResponse2005-print"></a>}}
\if{latex}{\out{\hypertarget{method-InlineResponse2005-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse2005$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InlineResponse2005-clone"></a>}}
\if{latex}{\out{\hypertarget{method-InlineResponse2005-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse2005$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
